mark FFI imports of Host<->Device memory transfers as safe
These also might take significant time, so allow the Haskell RTS to interrupt them. This is also the case for the memcpy*Async functions, as these are partially synchronous when given pageable host memory.

The overhead of the safe call was not noticeable on my machine, even for a 1 byte host->device transfer.