Add Cabal library as an explicit dependency
While not required to build the cuda package itself, Cabal >= 1.22 is required to run the Setup.hs script. Stating this dependency explicitly ensures that it gets installed. This is important when this package gets pulled in as a dependency, or for tools like stack where everything must be stated explicitly.

As far as I understand, this requirement on the Cabal library is different from the cabal-version field, which specifies the version of the syntax of the .cabal file. Thus, I think we can drop this back to the old version.