Add 'Describe' class for detailed descriptions of values
In particular we use this as a unified interface for describing things such as return status codes.

In particular this works around a problem where CUDA 8 removed a case from the CUcomputemode_enum. Since this is a read-only type used to describe the current state of the device, we continue deriving the corresponding Haskell data structure automatically and use this new class for cases where we want to describe what the compute mode actually means (the nvidia-device-query application).