Make Setup.hs backwards compatible with Cabal-1.22
This is an attempt to make the Setup script work with Cabal-1.22 and onwards. Doing so will make the package compile with out-of-the-box GHC-7.10. Note that the MIN_VERSION macros used to test which version of the Cabal library we are building against were only introduced in 1.24, so this is likely to still break for even older versions.