Record the path to the CUDA toolkit we compiled against
The bindings which are generated are somewhat dependent on which version of the CUDA toolkit it is compiled against. Furthermore, dependent packages might like to know, e.g., where nvcc is, or related libraries are located such as cufft and libnvvm, without searching again and hoping to arrive at the same place; linking against different versions will probably result in pain.